transfer then consume data the delete file

+++ src/main/java/exportDataExcel/dataExcel_api/config/camel/CamelFileImport.java
@@ -2,6 +2,7 @@ package exportDataExcel.dataExcel_api.config.camel;
 
 import org.apache.camel.builder.RouteBuilder;
 import org.springframework.batch.core.Job;
+import org.springframework.batch.core.JobExecution;
 import org.springframework.batch.core.JobParameters;
 import org.springframework.batch.core.JobParametersBuilder;
 import org.springframework.batch.core.launch.JobLauncher;
@@ -25,20 +26,33 @@ public class CamelFileImport extends RouteBuilder {
                             .addLong("startAt", System.currentTimeMillis())
                             .toJobParameters();
 
-                    jobLauncher.run(job, jobParameters);
+                    JobExecution jobExecution=jobLauncher.run(job, jobParameters);
+
+                    // Wait for the job to complete
+                    while (jobExecution.isRunning()) {
+                        System.out.println("Job is still running...");
+                        Thread.sleep(1000); // Check every second
+                    }
+
+                    // Check the job status
+                    if (jobExecution.getStatus().isUnsuccessful()) {
+                        System.out.println("Job failed with status: " + jobExecution.getStatus());
+                    } else {
+                        System.out.println("Job completed successfully with status: " + jobExecution.getStatus());
+
+                        // Proceed to delete the file only after the job completes successfully
+                        File file = new File("src/main/resources/data/camelOut/" + exchange.getIn().getHeader("CamelFileName"));
+                        if (file.exists()) {
+                            file.delete();
+                            System.out.println("File deleted: " + file.getName());
+                        }
+                    }
+
 
                 })
-                .log("Completed Process 0")
-                .to("direct:aggregate");
+                .log("Completed Process 0");
 
 
-        from("direct:process1")
-                .log("Starting Process 1").process(exchange -> {
-                    // After the batch process completes, delete the file from the output directory
-                    File file = new File("src/main/resources/data/camelOut/" + exchange.getIn().getHeader("CamelFileName"));
-                    if (file.exists()) {
-                        file.delete();
-                    }
-                });
+
     }
 }
